fix: 编译 warning 清理 + 唤醒词共享 models_ 防 double-free
warning 清理（3 处，零功能影响）:
- at_modem: 删 GetCsq 死代码 dbm
- axs5106l: ISR 声明去掉 IRAM_ATTR，消 section 冲突
- lcd_display: LVGL 枚举位运算显式 (int) 转换

唤醒词:
- afe/custom_wake_word 增 owns_models_ 标志，仅 deinit 自己 init 的
  models_；共享 models_list_ 由 audio_service 持有，回落 AFE 时不再
  double-deinit（防 UAF）
- mcp_server: 自定义唤醒词工具 + 入参校验

+++ main/display/lcd_display.cc
@@ -334,8 +334,8 @@ void LcdDisplay::SetupUI() {
     lv_obj_set_style_layout(status_bar_, LV_LAYOUT_NONE, 0);  // Use absolute positioning
     lv_obj_align(status_bar_, LV_ALIGN_TOP_MID, 0, 0);  // Overlap with top_bar_
     lv_obj_add_flag(status_bar_, LV_OBJ_FLAG_CLICKABLE);
-    lv_obj_set_style_bg_color(status_bar_, lv_color_hex(0xFFFFFF), (lv_style_selector_t)(LV_PART_MAIN | LV_STATE_PRESSED));
-    lv_obj_set_style_bg_opa(status_bar_, LV_OPA_20, (lv_style_selector_t)(LV_PART_MAIN | LV_STATE_PRESSED));
+    lv_obj_set_style_bg_color(status_bar_, lv_color_hex(0xFFFFFF), (lv_style_selector_t)((int)LV_PART_MAIN | (int)LV_STATE_PRESSED));
+    lv_obj_set_style_bg_opa(status_bar_, LV_OPA_20, (lv_style_selector_t)((int)LV_PART_MAIN | (int)LV_STATE_PRESSED));
 
     notification_label_ = lv_label_create(status_bar_);
     lv_obj_set_width(notification_label_, LV_HOR_RES * 0.8);
@@ -809,8 +809,8 @@ void LcdDisplay::SetupUI() {
     lv_obj_set_style_layout(status_bar_, LV_LAYOUT_NONE, 0);  // Use absolute positioning
     lv_obj_align(status_bar_, LV_ALIGN_TOP_MID, 0, 0);  // Overlap with top_bar_
     lv_obj_add_flag(status_bar_, LV_OBJ_FLAG_CLICKABLE);
-    lv_obj_set_style_bg_color(status_bar_, lv_color_hex(0xFFFFFF), (lv_style_selector_t)(LV_PART_MAIN | LV_STATE_PRESSED));
-    lv_obj_set_style_bg_opa(status_bar_, LV_OPA_20, (lv_style_selector_t)(LV_PART_MAIN | LV_STATE_PRESSED));
+    lv_obj_set_style_bg_color(status_bar_, lv_color_hex(0xFFFFFF), (lv_style_selector_t)((int)LV_PART_MAIN | (int)LV_STATE_PRESSED));
+    lv_obj_set_style_bg_opa(status_bar_, LV_OPA_20, (lv_style_selector_t)((int)LV_PART_MAIN | (int)LV_STATE_PRESSED));
 
     notification_label_ = lv_label_create(status_bar_);
     lv_obj_set_width(notification_label_, LV_HOR_RES * 0.75);
